Handwriting_Recognition_deep_learning/numberrecognize/requirements.txt
alabaster==0.7.8
anaconda-client==1.4.0
anaconda-navigator==1.2.1
argcomplete==1.0.0
astropy==1.2.1
Babel==2.3.3
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.0
boto==2.40.0
Bottleneck==1.1.0
certifi==2018.1.18
cffi==1.6.0
chardet==3.0.4
chest==0.2.3
click==6.6
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.7
comtypes==1.1.2
conda==4.3.22
conda-build==1.21.3
configobj==5.0.6
configparser==3.5.0
contextlib2==0.5.3
cryptography==1.4
cycler==0.10.0
Cython==0.24
cytoolz==0.8.0
dask==0.10.0
datashape==0.5.2
decorator==4.0.10
dill==0.2.5
dj-static==0.0.6
Django==1.9.6
docutils==0.12
dynd===c328ab7
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12.2
Flask-Cors==2.1.2
future==0.16.0
gevent==1.1.1
greenlet==0.4.10
h5py==2.6.0
HeapDict==1.0.0
html5lib==0.9999999
idna==2.6
imagesize==0.7.1
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==4.1.1
jupyter-core==4.1.0
Keras==2.0.5
keras-contrib==1.2.1
llvmlite==0.11.0
locket==0.2.0
lxml==3.6.0
Markdown==2.6.8
MarkupSafe==0.23
matplotlib==1.5.1
menuinst==1.4.1
mistune==0.7.2
mpmath==0.19
multipledispatch==0.4.8
nb-anacondacloud==1.1.0
nb-conda==1.1.0
nb-conda-kernels==1.0.3
nbconvert==4.2.0
nbformat==4.0.1
nbpresent==3.0.2
networkx==1.11
nibabel==2.1.0
nltk==3.2.1
nose==1.3.7
notebook==4.2.1
numba==0.26.0
numexpr==2.6.0
numpy==1.13.0
odo==0.5.0
opencv-python==3.2.0
openpyxl==2.3.2
pandas==0.18.1
partd==0.3.4
path.py==0.0.0
pathlib2==2.1.0
patsy==0.4.1
pep8==1.7.0
pickleshare==0.7.2
Pillow==3.2.0
ply==3.8
protobuf==3.3.0
psutil==4.3.0
py==1.4.31
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pydicom==0.9.9
pyexcel-io==0.2.3
pyexcel-xls==0.2.3
pyflakes==1.2.3
Pygments==2.1.3
PyMySQL==0.7.9
pyOpenSSL==16.2.0
pyparsing==2.1.4
pyreadline==2.1
pyserial==3.2.1
pytest==2.9.2
python-dateutil==2.5.3
pytz==2016.4
pywin32==220
PyYAML==3.12
pyzmq==15.2.0
qtconsole==4.2.1
QtPy==1.0.2
requests==2.18.4
rope-py3k==0.9.4.post1
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.17.1
simplegeneric==0.8.1
SimpleITK==0.10.0
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
spyder==2.3.9
SQLAlchemy==1.0.13
static3==0.7.0
statsmodels==0.6.1
sympy==1.0
tables==3.2.2
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3
Theano==0.9.0
toolz==0.8.0
tornado==4.3
traitlets==4.2.1
unicodecsv==0.14.1
urllib3==1.22
Werkzeug==0.12.2
xlrd==1.0.0
XlsxWriter==0.9.2
xlwings==0.7.2
xlwt==1.1.2
xlwt-future==0.8.0
